+++ tasks.py
@@ -1,6 +1,3 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals, absolute_import
-
 import os
 
 from invoke import task, call
@@ -85,16 +82,9 @@ def qa(ctx):
             error('There is some lints to fix')
         else:
             success('No lint to fix')
-        info('Ensure PyPI can render README and CHANGELOG')
-        readme_results = ctx.run('python setup.py check -r -s', pty=True, warn=True, hide=True)
-        if readme_results.failed:
-            print(readme_results.stdout)
-            error('README and/or CHANGELOG is not renderable by PyPI')
-        else:
-            success('README and CHANGELOG are renderable by PyPI')
-    if flake8_results.failed or readme_results.failed:
+    if flake8_results.failed:
         error('Quality check failed')
-        exit(flake8_results.return_code or readme_results.return_code)
+        exit(flake8_results.return_code)
     success('Quality check OK')
 
 
@@ -108,6 +98,7 @@ def dist(ctx, buildno=None):
     cmd.append('bdist_wheel')
     with ctx.cd(ROOT):
         ctx.run(' '.join(cmd), pty=True)
+    ctx.run('twine check dist/*')
     success('Distribution is available in dist directory')